+++ src/dungeon.h
@@ -1,6 +1,9 @@
 #ifndef DUNGEON_H
 #define DUNGEON_H
 
+struct Dungeon;
+typedef struct Dungeon Dungeon;
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdbool.h>
@@ -13,21 +16,23 @@
 #define DUNGEON_FLOORS_MIN 3
 #define DUNGEON_FLOORS_MAX 6
 
-typedef struct {
+struct Dungeon {
     u_char floorCount;
-
     u_char currentFloor;
 
     Player* player;
     Floor** floors;
     Settings* settings;
-} Dungeon;
+};
 
 Dungeon* dungeon_initialize(Settings* settings);
 Dungeon* dungeon_terminate(Dungeon* dungeon);
 int dungeon_load_from_program(Dungeon* dungeon);
-
 void dungeon_place_character(Dungeon* dungeon);
-void print_current_floor(Dungeon* dungeon);
+
+void dungeon_print_current_floor(Dungeon* dungeon);
+void dungeon_print_current_floor_hardness(Dungeon* dungeon);
+void dungeon_print_floor(Dungeon* dungeon, u_char floor);
+void dungeon_print_floor_hardness(Dungeon* dungeon, u_char floor);
 
 #endif