+++ src/escape_from_darkness.h
@@ -16,4 +16,6 @@
 int initialize(Dungeon** dungeon, int argc, char* argv[]);
 int terminate(Dungeon** dungeon);
 
+void game_tick(Dungeon* dungeon);
+
 #endif