+++ src/room.h
@@ -1,6 +1,9 @@
 #ifndef ROOM_H
 #define ROOM_H
 
+struct Room;
+typedef struct Room Room;
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdbool.h>
@@ -12,13 +15,13 @@
 #define ROOM_MIN_HEIGHT 4
 #define ROOM_MAX_HEIGHT 8
 
-typedef struct {
+struct Room{
     u_char width;
     u_char height;
 
     u_char startX;
     u_char startY;
-} Room;
+};
 
 Room* room_initialize(u_char startX, u_char startY, u_char width, u_char height);
 Room* room_terminate(Room* room);