+++ src/staircase.h
@@ -1,12 +1,15 @@
 #ifndef STAIRCASE_H
 #define STAIRCASE_H
 
+struct Staircase;
+typedef struct Staircase Staircase;
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdbool.h>
 #include "character_listings.h"
 
-typedef struct {
+struct Staircase {
     u_char x;
     u_char y;
 
@@ -15,7 +18,7 @@ typedef struct {
 
     bool isUp;
     bool isDown;
-} Staircase;
+};
 
 Staircase* staircase_initialize(u_char x, u_char y, u_char fromFloor, u_char toFloor);
 Staircase* staircase_terminate(Staircase* staircase);